0-Simulateur-Capteurs/requirements.txt
pandas
openpyxl
python-dotenv
kafka-python